+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/RemoteGitHub.java
@@ -27,7 +27,7 @@ public interface RemoteGitHub {
     UpdatedResource updateContent(String repoFullName, String path, DirectCommit directCommit, String oauthToken) throws
             GitHubAuthorizationException;
 
-    PullRequest createPullRequest(String repoFullName, PullRequestToCreate newPr);
+    PullRequest createPullRequest(String repoFullName, PullRequestToCreate newPr, String oauthToken)throws GitHubAuthorizationException;
 
     Repository fetchRepository(String repoFullName);